feat(repo-hygiene): move week2 root scripts into devtools with shims
