+++ docs/dev/LONG_TERM_REPO_STRATEGY.md
@@ -230,6 +230,10 @@ Delivery KPI:
   - 12주 로드맵의 단계별 DoD
   - 멀티 에이전트 운영 모델과 2주 작업지시서
   - dot 리소스(`.*`)의 유지/조건부/ignore 분류 원칙
+- Week 2 실행 반영:
+  - 루트 유틸 5개를 `scripts/devtools/`로 이관
+  - 루트 호환 shim 도입(점진 제거 전략)
+  - CI 빌드 경로 일부를 신규 위치로 전환
 
 ## 10) 요청 표준(Agent/Skill + PR 중심)
