+++ scripts/repo_hygiene_policy.json
@@ -62,6 +62,13 @@
       "run_tests.py",
       "setup_env.py"
     ],
+    "shim_file_names": [
+      "build_web_exe.py",
+      "build_web_exe_enhanced.py",
+      "cleanup_debug_files.py",
+      "fix_env_setup.py",
+      "run_tests.py"
+    ],
     "ignore_file_globs": [
       ".coverage",
       "coverage.xml",
